requirements.txt
chromadb
chroma-haystack
fastapi
haystack
llama-cpp-python
sentence-transformers